php-checks/src/main/java/org/sonar/php/checks/CheckBundle.java
/*
 * SonarQube PHP Plugin
 * Copyright (C) 2010-2023 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
package org.sonar.php.checks;

import java.util.List;
import org.sonar.plugins.php.api.visitors.CheckContext;
import org.sonar.plugins.php.api.visitors.PHPVisitorCheck;
import org.sonar.plugins.php.api.visitors.PhpIssue;

public abstract class CheckBundle extends PHPVisitorCheck {

  private List<CheckBundlePart> checks;

  @Override
  public void init() {
    checks = checks();
    checks.forEach(check -> {
      check.init();
      check.setBundle(this);
    });
  }

  @Override
  public List<PhpIssue> analyze(CheckContext context) {
    checks.forEach(check -> check.analyze(context));
    return context.getIssues();
  }

  protected abstract List<CheckBundlePart> checks();
}
